add Bash loadable extension, add check (unittest), add benchmark, add examples
A e2718 target is added for check and benchmark with the original e program.

Loadable is compiled as gnu99 and the gcc optimization for term function is set to -O0, otherwise, for input like -- -1 would result -nan. It can also be work around with the following in main function:

    printf("", p);

For some reason, GCC seems to optizmize it and ends up incorrect results.

-v option is added for loadable, which borrows from printf builtin.